+++ scripts/FollowerDetectionUtil/player.lua
@@ -1,10 +1,24 @@
 local followers = {}
 
+local function onSave()
+    return followers
+end
+
+local function onLoad(saveData)
+    if saveData then
+        followers = saveData
+    end
+end
+
 local function updateFollowerList(data)
     followers = data.followers
 end
 
 return {
+    engineHandlers = {
+        onSave = onSave,
+        onLoad = onLoad,
+    },
     eventHandlers = {
         FDU_UpdateFollowerList = updateFollowerList,
     },