+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Follower Detection Util (OpenMW)
 
+## 1.1
+
+### Changes
+
+- Follower list now also stores followers who don't follow player
+- All scripts now store full follower list in the save file
+- Minor optimizations all around
+
+### Fixes
+
+- Super leader field now stores actors instead of their states (for consistency)
+
 ## 1.0.1
 
 ### Fixes